backend/requirements.txt
# --- Core API Framework ---
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# --- AI & Execution ---
openai==1.14.3
langchain==0.1.13
pydantic-settings==2.2.1

# --- Database & Persistence ---
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1

# --- Security & Identity ---
# Switching to python-jose for better OAuth/Google compatibility
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
passlib[bcrypt]
email-validator>=2.0.0
# For Google OAuth and verification email triggers
httpx==0.27.0
emails==0.6

# --- Utilities ---
python-dotenv==1.0.1